+++ .travis.yml
@@ -19,7 +19,7 @@ before_script:
     - git clone --depth 1 https://github.com/apertium/apertium-nno /tmp/languages/apertium-nno || echo 'Using cached apertium-nno'
     - ( cd /tmp/languages/apertium-nno && ./autogen.sh && make -j2 )
 script:
-    - flake8 **/*.py
+    - flake8 *.py apertium_apy/ language_names/
     - if [[ $TRAVIS_PYTHON_VERSION != 3.4 ]]; then
         mypy --config-file mypy.ini **/*.py;
       fi
