+++ apertium_apy/__init__.py
@@ -1,0 +1,12 @@
+import random
+import string
+
+if False:
+    from typing import Optional  # noqa: F401
+
+    import apertium_apy.missingdb  # noqa: F401
+
+missing_freqs_db = None  # type: Optional[missingdb.MissingDb]  # has to be global for sig_handler :-/
+
+RECAPTCHA_VERIFICATION_URL = 'https://www.google.com/recaptcha/api/siteverify'
+BYPASS_TOKEN = ''.join(random.choice(string.ascii_letters + string.digits) for _ in range(24))
