+++ apertium_apy/keys.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 from collections import defaultdict
 
 if False:
