+++ apertium_apy/utils/wiki.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 import logging
 import json
