+++ tests/test.py
@@ -28,7 +28,7 @@ INSTALLEDPAIRS = os.environ.get('INSTALLEDPAIRS', '/usr/share/apertium')
 MAX_STARTUP_SECONDS = 10
 
 check_utf8()
-cli_args = shlex.split('-p {} -j1 -i3 -u1 -n1 -m3 -s "{}"  -- "{}"'.format(PORT, NONPAIRS, INSTALLEDPAIRS))
+cli_args = shlex.split('-p {} -v1 -j1 -i3 -u1 -n1 -m3 -s "{}"  -- "{}"'.format(PORT, NONPAIRS, INSTALLEDPAIRS))
 args = parse_args(cli_args=cli_args)
 enable_pretty_logging()
 application = setup_application(args)
@@ -93,7 +93,7 @@ class TestListHandler(BaseTestCase):
         self.assertIsNone(response['responseDetails'])
         self.assertEqual(response['responseStatus'], 200)
         expect = set(map(lambda x: frozenset(x.items()), [
-            {'sourceLanguage': 'sme', 'targetLanguage': 'nob'},
+            # {'sourceLanguage': 'sme', 'targetLanguage': 'nob'},
             {'sourceLanguage': 'eng', 'targetLanguage': 'spa'},
             {'sourceLanguage': 'spa', 'targetLanguage': 'eng_US'},
             {'sourceLanguage': 'spa', 'targetLanguage': 'eng'},
@@ -132,9 +132,9 @@ class TestTranslateHandler(BaseTestCase):
         response = self.fetch_translation('notaword', 'eng|spa', params={'markUnknown': False})
         self.assertEqual(response['responseData']['translatedText'], 'notaword')
 
-    def test_valid_giella_pair(self):
-        response = self.fetch_translation('ja', 'sme|nob')
-        self.assertEqual(response['responseData']['translatedText'], 'og')
+    # def test_valid_giella_pair(self):
+    #     response = self.fetch_translation('ja', 'sme|nob')
+    #     self.assertEqual(response['responseData']['translatedText'], 'og')
 
     def test_invalid_pair(self):
         response = self.fetch_translation('ignored', 'typomode', expect_success=False)
